+++ client/src/pages/Profile.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import StravaActivities from '../pages/strava'
+
+function Profile(){
+    return(
+        <div>
+            <StravaActivities />
+        </div>
+    )
+}
+  
+export default Profile;
