+++ client/src/pages/strava.js
@@ -5,18 +5,83 @@ import "../utils/strava";
 // import polyline from '@mapbox/polyline'
 
 function StravaActivities() {
-  const navigate = useNavigate();
+
+
+  // const navigate = useNavigate();
   const client_id = '77814';
   const params = useParams();
   console.log(window.location.search)
-  // const client_secret = 'ba4cf64706994d406df016b09df6d62ee55edaef';
-  // const refresh_token = 'e3c3bce7513bb09b9c19bfd2450855830fb0d313';
+  const client_secret = 'ba4cf64706994d406df016b09df6d62ee55edaef';
+
   useEffect(() => {
     console.log(params)
-    if(code){
+    if (code) {
+      function convertAuthToAccess() {
+        const authLink = `https://www.strava.com/oauth/token?client_id=${client_id}&client_secret=${client_secret}&code=${code}&grant_type3=authorization_code
+        `
+        console.log(authLink)
+      }
+      convertAuthToAccess();
       // add other function call to get user id
+
+      function getActivities(res) {
+        // commented out path to get user profile info from an array
+        // https://www.strava.com/api/v3/athlete?access_token=7bc114015e81808697585c211d65f269c319a3cc
+        const activitiesLink = `https://www.strava.com/api/v3/athlete/activities?access_token=${res.access_token}`
+
+        fetch(activitiesLink)
+          //   .then((res) => console.log(res.json()))
+          .then((res) => {
+            return res.json();
+          })
+          .then(data => {
+            console.log(data);
+            // var htmlContent = "";
+            // for (let i = 0; i > 5; i++) {
+            // document.getElementById('activity').innerHTML = data[0];
+            document.getElementById('activity-name').textContent = 'Activity Name: ' + data[0].name;
+
+            document.getElementById('activity-distance').innerHTML = 'Distance: ' + data[0].distance;
+
+            document.getElementById('averageSpeed').innerHTML = 'Average Speed: ' + data[0].average_speed;
+
+            document.getElementById('totalTime').innerHTML = 'Total Time: ' + data[0].elapsed_time;
+
+            document.getElementById('elevationGain').innerHTML = 'Total Elevation Gain: ' + data[0].total_elevation_gain;
+
+          })
+      }
+
+      //   getActivities();
+
+      //refresh token
+      //https://www.strava.com/oauth/token?client_id=77814&client_secret=ba4cf64706994d406df016b09df6d62ee55edaef&refresh_token=e3c3bce7513bb09b9c19bfd2450855830fb0d313&grant_type=refresh_token
+
+      const authLink = "https://www.strava.com/oauth/token";
+
+      function reAuthorize() {
+        fetch(authLink, {
+          method: 'post',
+          headers: {
+            'Accept': 'application/json, text/plain, */*',
+            'Content-Type': 'application/json'
+
+          },
+          body: JSON.stringify({
+            client_id: '77814',
+            client_secret: 'ba4cf64706994d406df016b09df6d62ee55edaef',
+            refresh_token: 'e3c3bce7513bb09b9c19bfd2450855830fb0d313',
+            grant_type: 'refresh_token'
+          })
+          //get response json it and pass it to activities
+        }).then((res => res.json()))
+          .then(res => getActivities(res))
+
+      }
+
+      reAuthorize();
     }
-    
+
   },
     [])
 
@@ -24,6 +89,7 @@ function StravaActivities() {
   // const term = url.get('term')
   console.log(url)
   // console.log(term)
+
   function getParameterByName(name, url = window.location.href) {
     name = name.replace(/[\[\]]/g, '\\$&');
     var regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)'),
@@ -32,11 +98,15 @@ function StravaActivities() {
     if (!results[2]) return '';
     return decodeURIComponent(results[2].replace(/\+/g, ' '));
   }
+
+
   var code = getParameterByName('code');
   console.log(code)
+
+
   function handleClick() {
-    window.location.href = `https://www.strava.com/oauth/authorize?client_id=${client_id}&redirect_uri=http://localhost:3000/social&response_type=code&scope=activity:read_all
-      `
+    window.location.href = `https://www.strava.com/oauth/authorize?client_id=${client_id}&redirect_uri=http://localhost:3000/profile&response_type=code&scope=activity:read_all
+        `
   }
   // function
